Move the fetchers classes
