+++ src/lightning/pytorch/loops/fetchers.py
